Organized test file + configuring coveralls
