+++ questions_factory/tests/__init__.py
@@ -1,0 +1,1 @@
+from tests_arithmetic_problem import *
